refactor(cli): extract runtime contracts
