+++ tests/cli/test_remote_release.py
@@ -7,7 +7,7 @@ from unittest.mock import MagicMock, patch
 
 from typer.testing import CliRunner
 
-from yoyopod.core.setup_contract import RUNTIME_REQUIRED_CONFIG_FILES
+from yoyopod_cli.contracts.setup import RUNTIME_REQUIRED_CONFIG_FILES
 from yoyopod_cli.remote_release import app as release_app
 from yoyopod_cli.slot_contract import (
     APP_NATIVE_RUNTIME_ARTIFACTS,
