+++ yoyopod_cli/__init__.py
@@ -1,8 +1,8 @@
-"""yoyopod_cli — flat CLI package for YoYoPod.
+"""yoyopod_cli - flat CLI package for YoYoPod.
 
 Entry point is `yoyopod_cli.main:run`. See COMMANDS.md for the full command reference.
 """
 
 from __future__ import annotations
 
-from yoyopod._version import __version__ as __version__
+from yoyopod_cli._version import __version__ as __version__
