+++ yoyopod_cli/contracts/__init__.py
@@ -1,7 +1,4 @@
-"""Compatibility re-export for setup contracts.
-
-Active CLI and deploy code should import from `yoyopod_cli.contracts.setup`.
-"""
+"""Production contracts shared by the Python CLI and deploy tooling."""
 
 from __future__ import annotations
